[feat] IssueContext를 IssuePage 컴포넌트로 끌어올림
IssueContext를 IssueSearchBox에서도 사용해야해서 IssuePage로 끌어올림.
