+++ frontend/src/components/IssueList/IssueList.js
@@ -1,26 +1,17 @@
-import React, { useEffect, useState, createContext } from 'react';
+import React, { useEffect, useState, useContext } from 'react';
 import { issueService } from '@services';
+import { IssueContext } from '@pages';
 import IssueFilterTab from './IssueFilterTab/IssueFilterTab';
 import IssueItem from './IssueItem/IssueItem';
 import { useHistory } from 'react-router-dom';
 
-const initialFilterOptions = {
-  isOpened: 1,
-  author: null,
-  label: [],
-  milestone: null,
-  assignee: null,
-};
-
-export const IssueContext = createContext();
-
 export default function IssueList() {
   const history = useHistory();
 
-  const [filterOptions, setFilterOptions] = useState(initialFilterOptions);
   const [issues, setIssues] = useState([]);
   const allChecked = issues.every(i => i.checked);
   const markMode = issues.some(i => i.checked);
+  const { filterOptions } = useContext(IssueContext);
 
   const setFilterdIssues = async options => {
     try {
@@ -44,7 +35,7 @@ export default function IssueList() {
   }, [filterOptions]);
 
   return (
-    <IssueContext.Provider value={{ filterOptions, setFilterOptions }}>
+    <>
       <IssueFilterTab
         issues={issues}
         setIssues={setIssues}
@@ -54,6 +45,6 @@ export default function IssueList() {
       {issues.map(({ no, title, checked }) => (
         <IssueItem key={no} title={title} checked={checked} issues={issues} setIssues={setIssues} />
       ))}
-    </IssueContext.Provider>
+    </>
   );
 }
