+++ frontend/src/components/index.js
@@ -1,7 +1,7 @@
 export { default as Header } from './Header/Header';
 export { default as LoginBox } from './LoginBox/LoginBox';
 export { default as SignUpBox } from './SignUpBox/SignUpBox';
-export { default as IssueList, IssueContext } from './IssueList/IssueList';
+export { default as IssueList } from './IssueList/IssueList';
 export { default as IssueSearchBox } from './IssueSearchBox/IssueSearchBox';
 export { default as LabelMilestoneControls } from './LabelMilestoneControls/LabelMilestoneControls';
 export { default as LabelBox } from './LabelBox/LabelBox';
