+++ frontend/src/pages/index.js
@@ -1,5 +1,5 @@
 export { default as NotFound } from './NotFound';
-export { default as Issue } from './Issue';
+export { default as Issue, IssueContext } from './Issue';
 export { default as Login } from './Login';
 export { default as SignUp } from './SignUp';
 export { default as Label } from './Label';
